# swagger.yaml
swagger: '2.0'
info:
  version: "1.0.0"
  title: Logistics Wizard Controller
  description: The main controller app for the Logistics Wizard system, interfacing with the ERP, Recommendations, and UI services.
  license:
    name: Apache 2.0
# the domain of the service
host: logistics-wizard.mybluemix.net
# array of all schemes that your API supports
schemes:
  - https
# will be prefixed to all paths
basePath: /api/v1
produces:
  - application/json
paths:
  '/users':
    get:
      summary: Get Users
      description: |
        Returns all users who have access to the Logistics Wizard system in some capacity.
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
      tags:
        - User
      responses:
        200:
          description: An array of users
          schema:
            type: array
            items:
              $ref: '#/definitions/User'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create User
      description: |
        Creates a new user in the system
      parameters:
        - name: user
          in: body
          description: User object
          required: true
          schema:
            $ref: '#/definitions/User'
      tags:
        - User
      responses:
        201:
          description: The new user object
          schema:
            $ref: '#/definitions/User'
        409:
          description: Username/email is taken
          schema:
            $ref: '#/definitions/Error'
        422:
          description: Invalid email
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/users/{id}':
    get:
      summary: Get User
      description: |
        Retrieves the user with the input user ID
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the user to be retrieved
          required: true
          type: string
      tags:
        - User
      responses:
        200:
          description: The retrieved User object
          schema:
            $ref: '#/definitions/User'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Update User
      description: |
        Updates attributes for a user
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the user to be retrieved
          required: true
          type: string
        - name: user
          in: body
          description: User object
          required: true
          schema:
            $ref: '#/definitions/User'
      tags:
        - User
      responses:
        200:
          description: The updated User object
          schema:
            $ref: '#/definitions/User'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/auth':
    post:
      summary: Sign In
      description: |
        Signs a specific user into the system and returns a bearer token
      parameters:
        - name: user
          in: body
          description: The User being signed in
          required: true
          schema:
            $ref: '#/definitions/User'
      tags:
        - User
      responses:
        200:
          description: The bearer token
          schema:
            $ref: '#/definitions/AccessToken'
        401:
          description: Username/password is incorrect
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/auth/{token}':
    delete:
      summary: User Logout
      description: |
        Logs out the user with the corresponding bearer token
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: token
          in: path
          description: Token of the user to be logged out
          required: true
          type: string
      tags:
        - User
      responses:
        204:
          description: Logout successful
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/shipments':
    get:
      summary: Retrieves Shipments
      description: |
        Retrieves a list of all shipments. Returned items will differ based on the identity and type of user logged in
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
      tags:
        - Shipment
      responses:
        200:
          description: An array of shipments
          schema:
            type: array
            items:
              $ref: '#/definitions/Shipment'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    post:
      summary: Create Shipment
      description: |
        Creates a new shipment in the ERP system
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: data
          in: body
          description: Shipment to be created
          required: true
          schema:
            $ref: '#/definitions/Shipment'
      tags:
        - Shipment
      responses:
        201:
          description: Shipment created successfully
          schema:
            $ref: '#/definitions/Shipment'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/shipments/{id}':
    get:
      summary: Retrieve shipment from the ERP
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the shipment to be retrieved
          required: true
          type: string
      tags:
        - Shipment
      responses:
        200:
          description: Shipment retrieved successfully
          schema:
            $ref: '#/definitions/Shipment'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    delete:
      summary: Delete a shipment from the ERP
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the shipment to be deleted
          required: true
          type: string
      tags:
        - Shipment
      responses:
        204:
          description: Shipment deleted successfully
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
    put:
      summary: Update attributes of a shipment in the ERP
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the shipment to be updated
          required: true
          type: string
        - name: data
          in: body
          description: The shipment object to replace the object currently associated with input ID
          required: false
          schema:
            $ref: '#/definitions/Shipment'
      tags:
        - Shipment
      responses:
        200:
          description: Shipment updated successfully
          schema:
            $ref: '#/definitions/Shipment'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/distribution-centers':
    get:
      summary: Retrieves Distribution Centers
      description: |
        Retrieves a list of distribution centers. Returned list will be dictated by user's access level
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
      tags:
        - DistributionCenter
      responses:
        200:
          description: Retrieved distribution centers successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/DistributionCenter'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/distribution-centers/{id}':
    get:
      summary: Retrieve distribution center from the ERP
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the distribution center to be retrieved
          required: true
          type: string
      tags:
        - DistributionCenter
      responses:
        200:
          description: Distribution center retrieved successfully
          schema:
            $ref: '#/definitions/DistributionCenter'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/retailers':
    get:
      summary: Retrieves Retailers
      description: |
        Retrieves a list of all retailers. Only accessible to supply chain admins
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
      tags:
        - Retailer
      responses:
        200:
          description: Retrieved retailers successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/Retailer'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/retailers/{id}':
    get:
      summary: Retrieve retailer from the ERP
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
        - name: id
          in: path
          description: ID of the retailer to be retrieved
          required: true
          type: string
      tags:
        - Retailer
      responses:
        200:
          description: Retailer retrieved successfully
          schema:
            $ref: '#/definitions/Retailer'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
  '/recommendations':
    get:
      summary: Retrieve Recommendations
      description: |
        Retrieves all recommendations relevant to the logged-in user
      parameters:
        - name: Authorization
          in: header
          description: Bearer token
          required: true
          type: string
      tags:
        - Recommendation
      responses:
        200:
          description: Retrieved recommendations successfully
          schema:
            type: array
            items:
              $ref: '#/definitions/Recommendation'
        401:
          description: Unauthorized
          schema:
            $ref: '#/definitions/Error'
        default:
          description: Unexpected error
          schema:
            $ref: '#/definitions/Error'
definitions:
  User:
    type: object
    properties:
      id:
        type: string
      email:
        type: string
      password:
        type: string
        format: password
      role:
        type: string
      createdAt:
        type: string
        format: date
    required:
      - email
      - password
  AccessToken:
    properties:
      id:
        type: string
      ttl:
        default: 1209600
        description: time to live in seconds (2 weeks by default)
        type: number
        format: double
      created:
        type: string
        format: date
      userId:
        type: string
    required:
      - id
    additionalProperties: false
  Product:
    properties:
      name:
        type: string
      id:
        type: string
      supplierId:
        type: string
    required:
      - name
  Supplier:
    properties:
      name:
        type: string
      id:
        type: string
    required:
      - name
    additionalProperties: false
  Shipment:
    properties:
      status:
        default: NEW
        type: string
      createdAt:
        type: string
        format: date
      updatedAt:
        type: string
        format: date
      deliveredAt:
        type: string
        format: date
      estimatedTimeOfArrival:
        type: string
        format: date
      id:
        type: string
      currentLocation:
        $ref: '#/definitions/Address'
      fromId:
        type: string
      toId:
        type: string
    required:
      - status
      - createdAt
    additionalProperties: false
  Inventory:
    properties:
      quantity:
        default: 0
        type: number
        format: double
      id:
        type: string
      productId:
        type: string
      dcId:
        type: string
    additionalProperties: false
  LineItem:
    properties:
      quantity:
        default: 0
        type: number
        format: double
      id:
        type: string
      shipmentId:
        type: string
    required:
      - quantity
    additionalProperties: false
  DistributionCenter:
    properties:
      id:
        type: string
      address:
        $ref: '#/definitions/Address'
      contact:
        $ref: '#/definitions/Contact'
    additionalProperties: false
  Retailer:
    properties:
      id:
        type: string
      address:
        $ref: '#/definitions/Address'
      contact:
        $ref: '#/definitions/Contact'
    additionalProperties: false
  Address:
    properties:
      city:
        type: string
      state:
        type: string
      country:
        type: string
      latitude:
        type: number
        format: double
      longitude:
        type: number
        format: double
      id:
        type: string
    additionalProperties: false
  Contact:
    properties:
      name:
        type: string
      id:
        type: string
    required:
      - name
    additionalProperties: false
  Recommendation:
    type: object
    properties:
      shipmentId:
        type: string
      justification:
        type: string
  Error:
    type: object
    properties:
      code:
        type: integer
        format: int32
      message:
        type: string
      fields:
        type: string
tags:
  - name: Supplier
  - name: Product
  - name: DistributionCenter
  - name: Shipment
  - name: Retailer
  - name: Inventory
  - name: User